requirements.txt
flask==3.0.3
waitress==3.0.0
transformers==4.43.3
torch==2.3.1
torchvision==0.18.1
python-dotenv==1.0.1
pytest==8.3.5
datasets==2.20.0
pillow==10.4.0
numpy==1.26.4
matplotlib==3.9.0